+++ src/pages/base/core/command/gitmoji/emoji.tsx
@@ -1,0 +1,22 @@
+import emoji from 'emojilib'
+
+console.log('emoji', emoji)
+
+const emojiList: any = []
+
+Object.entries(emoji).forEach(item => {
+	const icon = item[0]
+	const keys = item[1]
+	const _item = {
+		icon,
+		// @ts-ignore
+		key: keys.join(''),
+		desc: '',
+	}
+	emojiList.push(_item)
+})
+
+export const moji = {
+	type: 'Emoji',
+	list: emojiList,
+}