+++ src/pages/base/core/command/map.ts
@@ -1,5 +1,0 @@
-import { GITMOJI_COMMAND } from './gitmoji'
-
-export const keyMap = {
-	[GITMOJI_COMMAND.key]: GITMOJI_COMMAND,
-}