+++ src/pages/base/index.tsx
@@ -1,24 +1,15 @@
 import { useEffect, useState } from 'react'
-import {
-	Command,
-	CommandEmpty,
-	CommandGroup,
-	CommandInput,
-	CommandItem,
-	CommandList,
-	CommandSeparator,
-} from '@/components/ui/command'
+import { Command, CommandInput } from '@/components/ui/command'
 import { Smile } from 'lucide-react'
 import { Loading } from '@/components/common/Loading'
-import { copyToClipboard, inputFocus } from '@/lib/utils'
-import { toast } from 'sonner'
-import { catchEmojiAndCode } from './core/command/gitmoji/utils'
-import { gitMojiList } from './core/command/gitmoji/const'
+import { inputFocus } from '@/lib/utils'
+import { Selector } from '@/components/common/Selector'
+import { useEmojiSearch } from './core/core'
 
 const inputId = 'baseCommandInput'
 
 export const Base = () => {
-	const [inputText, setInputText] = useState<string>('')
+	const { filterText, setFilterText } = useEmojiSearch()
 
 	const [chooseMoji, setChooseMoji] = useState<string>('')
 
@@ -31,9 +22,6 @@ export const Base = () => {
 			<Command
 				onKeyDown={e => {
 					if (e.key === 'Enter') {
-						const { emoji } = catchEmojiAndCode(chooseMoji)
-						copyToClipboard(emoji)
-						toast('The copy has been copied', {})
 						setTimeout(() => {
 							inputFocus(inputId)
 						}, 100)
@@ -52,32 +40,36 @@ export const Base = () => {
 					icon={<Smile className="mr-2 shrink-0 opacity-50" />}
 					style={{ height: 60 }}
 					placeholder={'Please enter information to filter emoji'}
-					value={inputText}
+					value={filterText}
 					onInput={e => {
 						// @ts-ignore
-						setInputText(e.target.value)
+						setFilterText(e.target.value)
 					}}
 				/>
-				<CommandList>
+				<Selector />
+				{/* <CommandList>
 					<CommandEmpty>
 						<div className=" mt-4 w-full h-full flex flex-col justify-center items-center">
 							<Loading />
 							<div className="mt-4">No results found.</div>
 						</div>
 					</CommandEmpty>
-					<CommandGroup heading="Git">
-						{gitMojiList.map((item, index) => (
-							<CommandItem key={index}>
-								<div className=" flex items-center">
-									<div className=" text-3xl">{item.icon}</div>
-									<div className=" mx-2">{item.key}</div>
-									<div>{item.desc}</div>
-								</div>
-							</CommandItem>
-						))}
-					</CommandGroup>
+					{mojiList.map((item, index) => (
+						<CommandGroup heading={item.type} key={index}>
+							{item.list.map((_item, index) => (
+								<CommandItem key={index}>
+									<div className=" flex items-center">
+										<div className=" text-3xl">{_item.icon}</div>
+										<div className=" mx-2">{_item.key}</div>
+										<div>{_item.desc}</div>
+									</div>
+								</CommandItem>
+							))}
+						</CommandGroup>
+					))}
+
 					<CommandSeparator />
-				</CommandList>
+				</CommandList> */}
 			</Command>
 		</>
 	)