+++ CHANGELOG.md
@@ -4,7 +4,6 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### 0.0.2 (2024-12-12)
 
-
 ### Features
 
-* quick-emoji开发完成 ([7503d2f](https://github.com/Jimmylxue/qucik-emoji/commit/7503d2fbd841efaacb369eb880ccaa44adaa581c))
+- quick-emoji 开发完成 ([7503d2f](https://github.com/Jimmylxue/qucik-emoji/commit/7503d2fbd841efaacb369eb880ccaa44adaa581c))